requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-jwt-auth==0.5.0
filelock==3.12.2
h11==0.14.0
httptools==0.5.0
idna==3.4
platformdirs==3.7.0
pydantic==1.10.9
PyJWT==1.7.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
